requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
pyodbc==4.0.39
faker==19.3.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
flask==2.3.2